requirements.txt
google-cloud-bigquery==3.35.1
requests==2.32.5
functions-framework==3.*
python-dotenv